fs: Add IOCB_NOIO flag for generic_file_read_iter
upstream commit id: 41da51bce36f44eefc1e3d0f47d18841cbd065ba

Add an IOCB_NOIO flag that indicates to generic_file_read_iter that it
shouldn't trigger any filesystem I/O for the actual request or for
readahead.  This allows to do tentative reads out of the page cache as
some filesystems allow, and to take the appropriate locks and retry the
reads only if the requested pages are not cached.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Ni Xun <richardni@tencent.com>